Migrate to nullsafety

+++ lib/src/util/utils.dart
@@ -5,7 +5,7 @@ extension NumListExtension<T extends num> on Iterable<T> {
   T sumList() {
     if (T == int) {
       var sum = 0;
-      this.forEach((n) => sum += n);
+      this.forEach((n) => sum += n as int);
       return sum as T;
     } else if (T == double) {
       var sum = 0.0;
